+++ generator/internal/codec_sample/generate.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sample
+package codec_sample
 
 import (
 	"embed"
