+++ generator/internal/sidekick/refresh.go
@@ -19,13 +19,13 @@ import (
 	"path"
 
 	"github.com/googleapis/google-cloud-rust/generator/internal/api"
+	"github.com/googleapis/google-cloud-rust/generator/internal/codec_sample"
 	"github.com/googleapis/google-cloud-rust/generator/internal/config"
 	"github.com/googleapis/google-cloud-rust/generator/internal/dart"
 	"github.com/googleapis/google-cloud-rust/generator/internal/golang"
 	"github.com/googleapis/google-cloud-rust/generator/internal/parser"
 	"github.com/googleapis/google-cloud-rust/generator/internal/rust"
 	"github.com/googleapis/google-cloud-rust/generator/internal/rust_prost"
-	"github.com/googleapis/google-cloud-rust/generator/internal/sample"
 )
 
 func init() {
@@ -106,7 +106,7 @@ func refreshDir(rootConfig *config.Config, cmdLine *CommandLine, output string) 
 	case "dart":
 		return dart.Generate(model, output, config)
 	case "sample":
-		return sample.Generate(model, output, config)
+		return codec_sample.Generate(model, output, config)
 	default:
 		return fmt.Errorf("unknown language: %s", config.General.Language)
 	}
